Remove twine from production dependencies
Twine is not needed in production deployments, so was deleted from
`requirements.txt` (`-e .` will install everything in `setup.py`, which
does not contain `twine`). A `requirements-development.txt` was added,
containing `twine` and other libraries needed on the workflow to publish
to PyPI (so the workflow uses this new file instead of hard-coding every
requirement). The `python-jose` library was also removed from the
README, since it was not being used.