Add explicit override logic for exempted URIs in middleware
This will close and fix #27.

Using a simple `if` statement, adding one more check to see if the
processing of a given view should occur makes it simple to allow for
more complex behaviors on a `ModelViewSet` than either fully exempted or
exempted, yet able to be overriden in a View.